add tools.py
